refactor: split god modules into focused submodules
Break the four largest modules into smaller, single-purpose submodules,
keeping unit tests next to the code they exercise and using the modern
module layout (sibling files, no mod.rs):

- req-core/src/domain/tree.rs (1522 lines) -> tree.rs plus
  tree/{error,link,cycle,suspect}.rs: core storage and accessors stay in
  tree.rs; error types, link management, cycle detection, and
  fingerprint/HRID drift handling each get their own module.

- req-core/src/storage/directory.rs (1910 lines) -> directory.rs plus
  directory/{load,edit,maintenance,flush}.rs: the store and its
  read-only queries stay in directory.rs; loading from disk, mutations,
  drift/suspect-link maintenance, and flushing are split out. Public
  paths (storage::directory::{AcceptResult, AcceptSuspectLinkError,
  AddRequirementError, ...}) are preserved via re-exports.

- req-core/src/storage/markdown.rs (1002 lines) -> markdown.rs plus
  markdown/{frontmatter,content}.rs: YAML frontmatter types and the
  markdown body parser move out of the serialization core.

- req/src/cli/list.rs (1968 lines) -> list.rs plus
  list/{entry,filter,row,render}.rs, one module per pipeline stage:
  snapshotting entries, filtering, row production
  (traversal/sort/pagination), and output rendering. Shared test
  fixtures live in a cfg(test) fixtures module.

Also extracted along the way:

- Directory::add_requirement_with_namespace: inline title/body parsing
  moved to a split_title_and_body helper.
- List::run: row production moved to List::produce_rows and the header
  formatting to View::name/Filters::describe, dropping the
  too_many_lines allowance.
- req-mcp: fixed a nonexistent clippy lint name in an allow attribute
  (unused_async_trait_impl -> unused_async).
- Cargo.lock: pruned stale entries left behind by removed dependencies.

No behaviour changes; test count is unchanged (147) and all tests pass
apart from two pre-existing permission-based tests that cannot fail when
run as root.
