+++ req-core/src/domain/tree.rs
@@ -1,8 +1,16 @@
-//! New in-memory tree structure for requirements with decomposed storage
+//! In-memory tree structure for requirements with decomposed storage
 //!
 //! The [`Tree`] knows nothing about the filesystem or the directory structure.
 //! It stores requirements in a decomposed format for better maintainability and
 //! performance.
+//!
+//! The core storage and accessors live here; related behaviour is split into
+//! focused submodules:
+//!
+//! - `error`: error types for tree operations
+//! - `link`: creating, removing, and renaming parent-child links
+//! - `cycle`: cycle detection and prevention
+//! - `suspect`: fingerprint-based change detection and HRID drift repair
 
 use std::{
     collections::{BTreeMap, BTreeSet, HashMap, VecDeque},
@@ -10,7 +18,6 @@ use std::{
 };
 
 use petgraph::graphmap::DiGraphMap;
-use tracing::instrument;
 use uuid::Uuid;
 
 use crate::{
@@ -24,101 +31,14 @@ use crate::{
     Requirement,
 };
 
-/// Error type for tree insertion operations.
-#[derive(Debug)]
-pub enum TreeInsertError {
-    /// Attempted to insert a requirement with a UUID that already exists in the
-    /// tree.
-    DuplicateUuid {
-        /// The duplicate UUID
-        uuid: Uuid,
-    },
-    /// Attempted to insert a requirement with an HRID that already exists in
-    /// the tree.
-    DuplicateHrid {
-        /// The duplicate HRID
-        hrid: Hrid,
-        /// The UUID of the requirement being inserted
-        new_uuid: Uuid,
-        /// The UUID of the existing requirement with this HRID
-        existing_uuid: Uuid,
-    },
-}
+mod cycle;
+mod error;
+mod link;
+mod suspect;
 
-impl std::fmt::Display for TreeInsertError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::DuplicateUuid { uuid } => {
-                write!(f, "Duplicate requirement UUID: {uuid}")
-            }
-            Self::DuplicateHrid {
-                hrid,
-                new_uuid,
-                existing_uuid,
-            } => {
-                write!(
-                    f,
-                    "Duplicate requirement HRID: {} (attempting to insert UUID {}, but HRID \
-                     already maps to UUID {})",
-                    hrid.display(3),
-                    new_uuid,
-                    existing_uuid
-                )
-            }
-        }
-    }
-}
-
-impl std::error::Error for TreeInsertError {}
-
-/// Error type for accepting suspect links.
-#[derive(Debug, thiserror::Error)]
-pub enum AcceptLinkError {
-    /// The child requirement was not found in the tree.
-    #[error("Child requirement {0} not found")]
-    ChildNotFound(Uuid),
-
-    /// The parent requirement was not found in the tree.
-    #[error("Parent requirement {0} not found (may have failed to load or been deleted)")]
-    ParentNotFound(Uuid),
-
-    /// The link between child and parent does not exist.
-    #[error("No link exists between child {child} and parent {parent}")]
-    LinkNotFound {
-        /// The child UUID
-        child: Uuid,
-        /// The parent UUID
-        parent: Uuid,
-    },
-}
-
-/// Error type for linking requirements.
-#[derive(Debug, thiserror::Error)]
-pub enum LinkError {
-    /// The child requirement was not found in the tree.
-    #[error("Child UUID {0} not found in tree")]
-    ChildNotFound(Uuid),
-
-    /// The parent requirement was not found in the tree.
-    #[error("Parent UUID {0} not found in tree")]
-    ParentNotFound(Uuid),
-}
-
-/// Error type for linking requirements.
-#[derive(Debug, thiserror::Error)]
-pub enum LinkRequirementError {
-    /// The child HRID was not found in the tree.
-    #[error("Child requirement {0:?} not found")]
-    ChildNotFound(Hrid),
-
-    /// The parent HRID was not found in the tree.
-    #[error("Parent requirement {0:?} not found")]
-    ParentNotFound(Hrid),
-
-    /// The link would create a cycle in the requirement graph.
-    #[error("{0}")]
-    WouldCreateCycle(String),
-}
+pub use error::{AcceptLinkError, LinkError, LinkRequirementError, TreeInsertError};
+pub use link::LinkOutcome;
+pub use suspect::SuspectLink;
 
 /// Data stored on each edge in the dependency graph.
 ///
@@ -135,40 +55,6 @@ struct EdgeData {
     /// created. Used to detect if the parent has been modified since the
     /// link was established.
     fingerprint: String,
-}
-
-/// Vertex coloring state for cycle detection using the three-color DFS
-/// algorithm.
-///
-/// This implements a standard graph algorithm for detecting cycles in directed
-/// graphs. The three states track DFS traversal:
-///
-/// - **Gray**: The node is currently being visited (appears in the current
-///   recursion stack). A back edge to a Gray node indicates a cycle.
-/// - **Black**: The node has been completely processed (finished recursing on
-///   all descendants). A back edge to a Black node is a cross edge, not a
-///   cycle.
-/// - **Unvisited**: Nodes not in the map have never been visited.
-///
-/// # Algorithm
-///
-/// This is the standard three-color depth-first search algorithm:
-/// 1. Mark node Gray when entering (push to recursion stack)
-/// 2. Recursively visit unvisited neighbors
-/// 3. If we encounter a Gray ancestor, we've found a cycle (back edge)
-/// 4. Mark node Black when leaving (pop from recursion stack)
-///
-/// This allows us to detect cycles in a single O(V+E) pass while simultaneously
-/// extracting the complete cycle paths (e.g., `[A, B, C, A]`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum DfsColorForDetection {
-    /// Node is currently being visited (on the recursion stack).
-    /// A back edge to a Gray node indicates a cycle.
-    Gray,
-    /// Node has been completely finished. Edges to Black nodes are cross edges,
-    /// not cycles.
-    Black,
-    // Unvisited nodes are represented by the absence of an entry in the HashMap.
 }
 
 /// An in-memory representation of the set of requirements with decomposed
@@ -197,21 +83,6 @@ pub struct Tree {
     /// Edge data contains parent HRID and fingerprint for change detection.
     /// This is the sole source of truth for parent relationships.
     graph: DiGraphMap<Uuid, EdgeData>,
-}
-
-/// Result of linking two requirements together.
-#[derive(Debug)]
-pub struct LinkOutcome {
-    /// UUID of the child requirement.
-    pub child_uuid: Uuid,
-    /// HRID of the child requirement.
-    pub child_hrid: Hrid,
-    /// UUID of the parent requirement.
-    pub parent_uuid: Uuid,
-    /// HRID of the parent requirement.
-    pub parent_hrid: Hrid,
-    /// Whether the relationship already existed prior to linking.
-    pub already_linked: bool,
 }
 
 impl Default for Tree {
@@ -472,152 +343,6 @@ impl Tree {
         Ok(())
     }
 
-    /// Link two requirements identified by their HRIDs.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error when either HRID does not exist in the tree, when
-    /// the parent/child UUIDs cannot be linked, or when the link would create a
-    /// cycle.
-    pub fn link_requirement(
-        &mut self,
-        child: &Hrid,
-        parent: &Hrid,
-    ) -> Result<LinkOutcome, LinkRequirementError> {
-        let (child_uuid, child_hrid) = {
-            let view = self
-                .find_by_hrid(child)
-                .ok_or_else(|| LinkRequirementError::ChildNotFound(child.clone()))?;
-            (*view.uuid, view.hrid.clone())
-        };
-
-        let (parent_uuid, parent_hrid, parent_fingerprint) = {
-            let view = self
-                .find_by_hrid(parent)
-                .ok_or_else(|| LinkRequirementError::ParentNotFound(parent.clone()))?;
-            (*view.uuid, view.hrid.clone(), view.fingerprint())
-        };
-
-        // Check if this link would create a cycle
-        self.check_would_create_cycle(child_uuid, parent_uuid)
-            .map_err(|e| LinkRequirementError::WouldCreateCycle(e.to_string()))?;
-
-        let already_linked = self.graph.contains_edge(child_uuid, parent_uuid);
-
-        self.upsert_parent_link(child_uuid, parent_uuid, parent_fingerprint)
-            .map_err(|error| match error {
-                LinkError::ChildNotFound(_) => {
-                    LinkRequirementError::ChildNotFound(child_hrid.clone())
-                }
-                LinkError::ParentNotFound(_) => {
-                    LinkRequirementError::ParentNotFound(parent_hrid.clone())
-                }
-            })?;
-
-        Ok(LinkOutcome {
-            child_uuid,
-            child_hrid,
-            parent_uuid,
-            parent_hrid,
-            already_linked,
-        })
-    }
-
-    /// Unlink two requirements identified by their HRIDs.
-    ///
-    /// Removes the parent-child relationship between the two requirements.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error when:
-    /// - Either HRID does not exist in the tree
-    /// - The link between child and parent does not exist
-    pub fn unlink_requirement(&mut self, child: &Hrid, parent: &Hrid) -> anyhow::Result<Uuid> {
-        let child_uuid =
-            self.hrid_to_uuid.get(child).copied().ok_or_else(|| {
-                anyhow::anyhow!("Child requirement {} not found", child.display(3))
-            })?;
-
-        let parent_uuid =
-            self.hrid_to_uuid.get(parent).copied().ok_or_else(|| {
-                anyhow::anyhow!("Parent requirement {} not found", parent.display(3))
-            })?;
-
-        // Check if the edge exists
-        if !self.graph.contains_edge(child_uuid, parent_uuid) {
-            anyhow::bail!(
-                "No link exists between {} and {}",
-                child.display(3),
-                parent.display(3)
-            );
-        }
-
-        // Remove the edge
-        self.graph.remove_edge(child_uuid, parent_uuid);
-
-        // Return child UUID so Directory can mark it dirty
-        Ok(child_uuid)
-    }
-
-    /// Renames a requirement by changing its HRID.
-    ///
-    /// This updates:
-    /// - The HRID mapping for the requirement
-    /// - All parent edges that reference this requirement (so children's parent
-    ///   HRIDs are updated)
-    ///
-    /// Returns the UUID of the renamed requirement and a list of children UUIDs
-    /// (which need to be marked dirty since their parent HRID changed).
-    ///
-    /// # Errors
-    ///
-    /// Returns an error when:
-    /// - The old HRID does not exist
-    /// - The new HRID already exists
-    /// - The new HRID is invalid for the current configuration
-    pub fn rename_requirement(
-        &mut self,
-        old_hrid: &Hrid,
-        new_hrid: &Hrid,
-    ) -> anyhow::Result<(Uuid, Vec<Uuid>)> {
-        // Check old HRID exists
-        let uuid = self
-            .hrid_to_uuid
-            .get(old_hrid)
-            .copied()
-            .ok_or_else(|| anyhow::anyhow!("Requirement {} not found", old_hrid.display(3)))?;
-
-        // Check new HRID doesn't exist
-        if self.hrid_to_uuid.contains_key(new_hrid) {
-            anyhow::bail!(
-                "Cannot rename to {}: HRID already exists",
-                new_hrid.display(3)
-            );
-        }
-
-        // Update HRID mappings
-        self.hrids.insert(uuid, new_hrid.clone());
-        self.hrid_to_uuid.remove(old_hrid);
-        self.hrid_to_uuid.insert(new_hrid.clone(), uuid);
-
-        // Find all children (incoming edges where this requirement is the parent)
-        let children: Vec<Uuid> = self
-            .graph
-            .edges_directed(uuid, petgraph::Direction::Incoming)
-            .map(|(child_uuid, _, _)| child_uuid)
-            .collect();
-
-        // Update all edges where this requirement is referenced as a parent
-        // We need to update the EdgeData to have the new parent HRID
-        for child_uuid in &children {
-            if let Some(edge_data) = self.graph.edge_weight_mut(*child_uuid, uuid) {
-                edge_data.parent_hrid = new_hrid.clone();
-            }
-        }
-
-        Ok((uuid, children))
-    }
-
     /// Get all children of a requirement.
     #[must_use]
     pub fn children(&self, uuid: Uuid) -> Vec<Uuid> {
@@ -643,133 +368,6 @@ impl Tree {
             .edges(uuid)
             .map(|(_, parent_uuid, edge_data)| (parent_uuid, edge_data.fingerprint.clone()))
             .collect()
-    }
-
-    /// Insert or update a parent link for the given child UUID.
-    ///
-    /// Returns `Ok(true)` if an existing link was replaced, or `Ok(false)` if a
-    /// new link was created.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if either the child or parent UUID does not exist in
-    /// the tree.
-    pub fn upsert_parent_link(
-        &mut self,
-        child_uuid: Uuid,
-        parent_uuid: Uuid,
-        fingerprint: String,
-    ) -> Result<bool, LinkError> {
-        // Validate both UUIDs exist in requirements first
-        if !self.requirements.contains_key(&child_uuid) {
-            return Err(LinkError::ChildNotFound(child_uuid));
-        }
-        if !self.requirements.contains_key(&parent_uuid) {
-            return Err(LinkError::ParentNotFound(parent_uuid));
-        }
-
-        // Ensure both nodes exist in the graph; GraphMap::add_node is idempotent.
-        self.graph.add_node(child_uuid);
-        self.graph.add_node(parent_uuid);
-
-        let parent_hrid = self
-            .hrids
-            .get(&parent_uuid)
-            .ok_or(LinkError::ParentNotFound(parent_uuid))?;
-
-        let edge = EdgeData {
-            parent_hrid: parent_hrid.clone(),
-            fingerprint,
-        };
-
-        Ok(self.graph.add_edge(child_uuid, parent_uuid, edge).is_some())
-    }
-
-    /// Check which requirements have stale parent HRIDs without modifying them.
-    ///
-    /// Returns an iterator of child UUIDs that have at least one parent link
-    /// with an outdated HRID.
-    #[instrument(skip(self))]
-    pub fn check_hrid_drift(&self) -> impl Iterator<Item = Uuid> + '_ {
-        use std::collections::HashSet;
-
-        let mut drifted_uuids = HashSet::new();
-
-        for child_uuid in self.graph.nodes() {
-            for (_, parent_uuid, edge_data) in self.graph.edges(child_uuid) {
-                // Get the current HRID of the parent
-                let Some(current_parent_hrid) = self.hrids.get(&parent_uuid) else {
-                    continue;
-                };
-
-                // Check if the stored HRID is outdated
-                if &edge_data.parent_hrid != current_parent_hrid {
-                    drifted_uuids.insert(child_uuid);
-                }
-            }
-        }
-
-        drifted_uuids.into_iter()
-    }
-
-    /// Update parent HRIDs in all requirements.
-    ///
-    /// When requirements are renamed or moved, the stored parent HRIDs in child
-    /// requirements can become stale. This method updates all outdated parent
-    /// HRIDs to match their current values.
-    ///
-    /// Returns an iterator of child UUIDs that were updated.
-    ///
-    /// Edges whose parent is missing from the tree (e.g. failed to load) are
-    /// skipped with a warning.
-    #[instrument(skip(self))]
-    pub fn update_hrids(&mut self) -> impl Iterator<Item = Uuid> + '_ {
-        use std::collections::HashSet;
-
-        let mut updated_uuids = HashSet::new();
-
-        // Collect all edges that need updating (store edge identifiers, not cloned
-        // HRIDs)
-        let mut edges_to_update = Vec::new();
-
-        for child_uuid in self.graph.nodes() {
-            for (_, parent_uuid, edge_data) in self.graph.edges(child_uuid) {
-                // Get the current HRID of the parent
-                // Skip edges with missing parents (e.g., failed to load)
-                let Some(current_parent_hrid) = self.hrids.get(&parent_uuid) else {
-                    tracing::warn!(
-                        "Skipping edge with missing parent: child={child_uuid}, \
-                         parent={parent_uuid}"
-                    );
-                    continue;
-                };
-
-                // Check if the stored HRID is outdated
-                if &edge_data.parent_hrid != current_parent_hrid {
-                    edges_to_update.push((child_uuid, parent_uuid));
-                    updated_uuids.insert(child_uuid);
-                }
-            }
-        }
-
-        // Apply the updates to EdgeData only
-        for (child_uuid, parent_uuid) in edges_to_update {
-            // Look up the HRID again (HashMap lookup is O(1) and cheaper than cloning
-            // earlier). Should always succeed since we already checked above.
-            let Some(current_parent_hrid) = self.hrids.get(&parent_uuid) else {
-                // Parent disappeared between collection and update - skip
-                tracing::warn!(
-                    "Parent {parent_uuid} disappeared during HRID update for child {child_uuid}"
-                );
-                continue;
-            };
-            if let Some(edge_data) = self.graph.edge_weight_mut(child_uuid, parent_uuid) {
-                // Update EdgeData (the sole source of truth)
-                edge_data.parent_hrid = current_parent_hrid.clone();
-            }
-        }
-
-        updated_uuids.into_iter()
     }
 
     /// Return all ancestors (transitive parents) of a requirement as UUIDs.
@@ -830,693 +428,5 @@ impl Tree {
         }
 
         visited.into_iter().collect()
-    }
-
-    /// Find all suspect links in the requirement graph.
-    ///
-    /// A link is suspect when the fingerprint stored in the edge data
-    /// does not match the current fingerprint of the parent requirement.
-    ///
-    /// Graph nodes without a corresponding HRID (e.g. failed to load) are
-    /// skipped.
-    #[must_use]
-    pub fn suspect_links(&self) -> Vec<SuspectLink> {
-        use crate::domain::requirement::ContentRef;
-
-        let mut suspect = Vec::new();
-
-        for child_uuid in self.graph.nodes() {
-            let Some(child_hrid) = self.hrids.get(&child_uuid) else {
-                // Skip nodes that failed to register an HRID instead of panicking.
-                continue;
-            };
-
-            for (_, parent_uuid, edge_data) in self.graph.edges(child_uuid) {
-                // Access RequirementData directly to avoid full RequirementView construction
-                let parent_data = self.requirements.get(&parent_uuid);
-
-                // Calculate current fingerprint, or use empty string if parent is missing
-                // Empty string indicates a dangling/broken reference (parent failed to load or
-                // was deleted)
-                let current_fingerprint = parent_data.map_or_else(String::new, |data| {
-                    ContentRef {
-                        title: &data.title,
-                        body: &data.body,
-                        tags: &data.tags,
-                    }
-                    .fingerprint()
-                });
-
-                // Report as suspect if fingerprints don't match, OR if parent is missing
-                if edge_data.fingerprint != current_fingerprint {
-                    suspect.push(SuspectLink {
-                        child_uuid,
-                        child_hrid: child_hrid.clone(),
-                        parent_uuid,
-                        parent_hrid: edge_data.parent_hrid.clone(),
-                        stored_fingerprint: edge_data.fingerprint.clone(),
-                        current_fingerprint,
-                    });
-                }
-            }
-        }
-
-        suspect
-    }
-
-    /// Update the fingerprint for a specific parent link.
-    ///
-    /// Returns `Ok(true)` if the fingerprint was updated, `Ok(false)` if
-    /// already up to date.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the child or parent requirement is not found, or if
-    /// the link doesn't exist.
-    pub fn accept_suspect_link(
-        &mut self,
-        child_uuid: Uuid,
-        parent_uuid: Uuid,
-    ) -> Result<bool, AcceptLinkError> {
-        use crate::domain::requirement::ContentRef;
-
-        // Check if child exists in graph
-        if !self.graph.contains_node(child_uuid) {
-            return Err(AcceptLinkError::ChildNotFound(child_uuid));
-        }
-
-        // Compute the parent's current fingerprint directly from its stored
-        // data; building a full RequirementView here would clone every parent
-        // and child link just to hash the content.
-        let current_fingerprint = self
-            .requirements
-            .get(&parent_uuid)
-            .map(|data| {
-                ContentRef {
-                    title: &data.title,
-                    body: &data.body,
-                    tags: &data.tags,
-                }
-                .fingerprint()
-            })
-            .ok_or(AcceptLinkError::ParentNotFound(parent_uuid))?;
-
-        // Find and update the edge
-        if let Some(edge_data) = self.graph.edge_weight_mut(child_uuid, parent_uuid) {
-            if edge_data.fingerprint == current_fingerprint {
-                return Ok(false); // Already up to date
-            }
-
-            // Update EdgeData (the sole source of truth)
-            edge_data.fingerprint.clone_from(&current_fingerprint);
-
-            Ok(true)
-        } else {
-            Err(AcceptLinkError::LinkNotFound {
-                child: child_uuid,
-                parent: parent_uuid,
-            })
-        }
-    }
-
-    /// Update all suspect fingerprints in the tree.
-    ///
-    /// Skips links where the parent is missing (logs a warning instead of
-    /// failing).
-    pub fn accept_all_suspect_links(&mut self) -> Vec<(Uuid, Uuid)> {
-        let suspect = self.suspect_links();
-        let mut updated = Vec::new();
-
-        for link in suspect {
-            match self.accept_suspect_link(link.child_uuid, link.parent_uuid) {
-                Ok(true) => {
-                    updated.push((link.child_uuid, link.parent_uuid));
-                }
-                Ok(false) => {
-                    // Already up to date, skip
-                }
-                Err(AcceptLinkError::ParentNotFound(parent_uuid)) => {
-                    // Parent missing - log but don't fail the entire operation
-                    tracing::warn!(
-                        "Cannot accept suspect link: parent {parent_uuid} not found (child: {})",
-                        link.child_uuid
-                    );
-                }
-                Err(e) => {
-                    // Other errors are unexpected but also shouldn't stop the batch
-                    tracing::error!("Failed to accept suspect link: {e}");
-                }
-            }
-        }
-
-        updated
-    }
-
-    /// Detect all cycles in the requirement graph using three-color depth-first
-    /// search.
-    ///
-    /// Returns a list of cycles, where each cycle is represented as a path of
-    /// HRIDs that forms a loop (e.g., `vec![USR-001, SYS-002, USR-001]`).
-    ///
-    /// # Algorithm
-    ///
-    /// Uses the standard three-color DFS algorithm to detect back edges in
-    /// O(V+E) time. When a back edge to a Gray ancestor is found, the full
-    /// cycle path is extracted and deduplicated before being added to the
-    /// result.
-    ///
-    /// # Example
-    ///
-    /// If the graph has edges: `A→B→C→A`, this returns `vec![vec![A, B, C, A]]`
-    #[must_use]
-    pub fn detect_cycles(&self) -> Vec<Vec<Hrid>> {
-        use std::collections::HashMap;
-
-        let mut colors: HashMap<Uuid, DfsColorForDetection> = HashMap::new();
-        let mut cycles = Vec::new();
-
-        // Start DFS from each unvisited node to ensure we find all cycles
-        for start_node in self.graph.nodes() {
-            if !colors.contains_key(&start_node) {
-                self.dfs_detect_cycles(start_node, &mut colors, &mut cycles);
-            }
-        }
-
-        cycles
-    }
-
-    /// Iterative helper for three-color DFS cycle detection.
-    ///
-    /// Each stack frame holds a node, its outgoing neighbours, and the index
-    /// of the next neighbour to visit, so deep parent chains cannot overflow
-    /// the call stack. The current DFS path is maintained to extract full
-    /// cycle information when back edges (edges to Gray nodes) are found.
-    fn dfs_detect_cycles(
-        &self,
-        start: Uuid,
-        colors: &mut std::collections::HashMap<Uuid, DfsColorForDetection>,
-        cycles: &mut Vec<Vec<Hrid>>,
-    ) {
-        use self::DfsColorForDetection::{Black, Gray};
-
-        let neighbours_of =
-            |node: Uuid| -> Vec<Uuid> { self.graph.edges(node).map(|(_, p, _)| p).collect() };
-
-        let mut path: Vec<Uuid> = Vec::new();
-        let mut stack: Vec<(Uuid, Vec<Uuid>, usize)> = Vec::new();
-
-        colors.insert(start, Gray);
-        path.push(start);
-        stack.push((start, neighbours_of(start), 0));
-
-        while let Some(frame) = stack.last_mut() {
-            let node = frame.0;
-            let next = frame.1.get(frame.2).copied();
-            frame.2 += 1;
-
-            let Some(parent_uuid) = next else {
-                // All outgoing edges visited: mark Black and leave the node.
-                colors.insert(node, Black);
-                path.pop();
-                stack.pop();
-                continue;
-            };
-
-            match colors.get(&parent_uuid) {
-                Some(Gray) => {
-                    // Back edge found! This node is an ancestor in the current path.
-                    // Extract and record the cycle.
-                    if let Some(pos) = path.iter().position(|&u| u == parent_uuid) {
-                        let cycle_path: Vec<Hrid> = path[pos..]
-                            .iter()
-                            .chain(std::iter::once(&parent_uuid))
-                            .filter_map(|&uuid| self.hrids.get(&uuid).cloned())
-                            .collect();
-                        if !cycle_path.is_empty()
-                            && !cycles.iter().any(|c| Self::cycles_equal(c, &cycle_path))
-                        {
-                            cycles.push(cycle_path);
-                        }
-                    }
-                }
-                Some(Black) => {
-                    // Already processed this subtree, skip (cross edge)
-                }
-                None => {
-                    // Unvisited node, descend
-                    colors.insert(parent_uuid, Gray);
-                    path.push(parent_uuid);
-                    stack.push((parent_uuid, neighbours_of(parent_uuid), 0));
-                }
-            }
-        }
-    }
-
-    /// Check if two cycles are equivalent (same HRIDs, possibly rotated).
-    ///
-    /// Cycle paths are closed walks (`[A, B, C, A]`): the comparison drops the
-    /// closing element that repeats the start, then tests whether one sequence
-    /// is a rotation of the other.
-    fn cycles_equal(a: &[Hrid], b: &[Hrid]) -> bool {
-        fn open(cycle: &[Hrid]) -> &[Hrid] {
-            match cycle {
-                [rest @ .., last] if rest.first() == Some(last) => rest,
-                _ => cycle,
-            }
-        }
-
-        let a = open(a);
-        let b = open(b);
-        if a.len() != b.len() {
-            return false;
-        }
-        a.is_empty()
-            || (0..b.len()).any(|offset| (0..a.len()).all(|i| a[i] == b[(offset + i) % b.len()]))
-    }
-
-    /// Check if adding a link from child to parent would create a cycle.
-    ///
-    /// Returns `Ok(())` if the link is safe, or `Err()` with the cycle path
-    /// that would be created.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if creating the link would form a cycle in the graph.
-    pub fn check_would_create_cycle(
-        &self,
-        child_uuid: Uuid,
-        parent_uuid: Uuid,
-    ) -> anyhow::Result<()> {
-        // If the parent can reach the child in the graph, then child -> parent would
-        // create a cycle
-        if self.can_reach(parent_uuid, child_uuid) {
-            // Find the cycle path for error reporting
-            let cycle_path = self.find_cycle_path(parent_uuid, child_uuid);
-            let cycle_str = cycle_path
-                .iter()
-                .filter_map(|&uuid| self.hrids.get(&uuid).map(|h| format!("{}", h.display(3))))
-                .collect::<Vec<_>>()
-                .join(" → ");
-
-            let child_hrid = self.hrids.get(&child_uuid).map_or_else(
-                || format!("(UUID: {child_uuid})"),
-                |h| format!("{}", h.display(3)),
-            );
-
-            anyhow::bail!("Cannot create link: would form a cycle: {cycle_str} → {child_hrid}");
-        }
-        Ok(())
-    }
-
-    /// Check if there's a path from source to target in the graph.
-    fn can_reach(&self, source: Uuid, target: Uuid) -> bool {
-        use std::collections::{HashSet, VecDeque};
-
-        if !self.graph.contains_node(source) {
-            return false;
-        }
-
-        let mut visited = HashSet::new();
-        let mut queue = VecDeque::new();
-        queue.push_back(source);
-        visited.insert(source);
-
-        while let Some(node) = queue.pop_front() {
-            if node == target {
-                return true;
-            }
-
-            for (_, parent_uuid, _) in self.graph.edges(node) {
-                if visited.insert(parent_uuid) {
-                    queue.push_back(parent_uuid);
-                }
-            }
-        }
-
-        false
-    }
-
-    /// Find a path from source to target for cycle reporting.
-    fn find_cycle_path(&self, source: Uuid, target: Uuid) -> Vec<Uuid> {
-        use std::collections::{HashMap, VecDeque};
-
-        let mut parent_map: HashMap<Uuid, Uuid> = HashMap::new();
-        let mut queue = VecDeque::new();
-        queue.push_back(source);
-        parent_map.insert(source, source);
-
-        while let Some(node) = queue.pop_front() {
-            if node == target {
-                // Reconstruct path
-                let mut path = vec![target];
-                let mut current = target;
-                while current != source {
-                    current = parent_map[&current];
-                    path.push(current);
-                }
-                path.reverse();
-                return path;
-            }
-
-            for (_, parent_uuid, _) in self.graph.edges(node) {
-                if let std::collections::hash_map::Entry::Vacant(e) = parent_map.entry(parent_uuid)
-                {
-                    e.insert(node);
-                    queue.push_back(parent_uuid);
-                }
-            }
-        }
-
-        vec![]
-    }
-}
-
-/// A suspect link in the requirement graph.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SuspectLink {
-    /// The UUID of the child requirement.
-    pub child_uuid: Uuid,
-    /// The HRID of the child requirement.
-    pub child_hrid: Hrid,
-    /// The UUID of the parent requirement.
-    pub parent_uuid: Uuid,
-    /// The HRID of the parent requirement.
-    pub parent_hrid: Hrid,
-    /// The fingerprint stored in the child's parent reference.
-    pub stored_fingerprint: String,
-    /// The current fingerprint of the parent requirement.
-    ///
-    /// If empty, indicates the parent requirement is missing (failed to load or
-    /// was deleted).
-    pub current_fingerprint: String,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::domain::Requirement;
-
-    #[test]
-    fn test_upsert_parent_link_errors_on_missing_parent() {
-        // Test that upsert_parent_link returns an error (not panic) when
-        // trying to link to a parent that doesn't exist
-        let mut tree = Tree::default();
-
-        // Create and insert a child requirement
-        let child_req = Requirement::new(
-            "USR-001".parse().unwrap(),
-            "Child requirement".to_string(),
-            String::new(),
-        );
-        let child_uuid = child_req.uuid();
-        tree.insert(child_req).unwrap();
-
-        // Try to link to a parent UUID that doesn't exist in the tree
-        let missing_parent_uuid = uuid::Uuid::new_v4();
-
-        // This should return an error, not panic
-        let result =
-            tree.upsert_parent_link(child_uuid, missing_parent_uuid, "fingerprint".to_string());
-
-        // Verify we get a ParentNotFound error
-        assert!(result.is_err(), "Expected error for missing parent");
-        match result {
-            Err(LinkError::ParentNotFound(uuid)) => {
-                assert_eq!(
-                    uuid, missing_parent_uuid,
-                    "Error should contain the missing parent UUID"
-                );
-            }
-            _ => panic!("Expected ParentNotFound error, got: {result:?}"),
-        }
-    }
-
-    #[test]
-    fn test_upsert_parent_link_errors_on_missing_child() {
-        // Test that upsert_parent_link returns an error when child doesn't exist
-        let mut tree = Tree::default();
-
-        // Create and insert a parent requirement
-        let parent_req = Requirement::new(
-            "SYS-001".parse().unwrap(),
-            "Parent requirement".to_string(),
-            String::new(),
-        );
-        let parent_uuid = parent_req.uuid();
-        tree.insert(parent_req).unwrap();
-
-        // Try to link a child UUID that doesn't exist
-        let missing_child_uuid = uuid::Uuid::new_v4();
-
-        // This should return an error, not panic
-        let result =
-            tree.upsert_parent_link(missing_child_uuid, parent_uuid, "fingerprint".to_string());
-
-        // Verify we get a ChildNotFound error
-        assert!(result.is_err(), "Expected error for missing child");
-        match result {
-            Err(LinkError::ChildNotFound(uuid)) => {
-                assert_eq!(
-                    uuid, missing_child_uuid,
-                    "Error should contain the missing child UUID"
-                );
-            }
-            _ => panic!("Expected ChildNotFound error, got: {result:?}"),
-        }
-    }
-
-    #[test]
-    fn test_upsert_parent_link_succeeds_when_both_exist() {
-        // Test the happy path - both parent and child exist
-        let mut tree = Tree::default();
-
-        // Create and insert parent and child requirements
-        let parent_req = Requirement::new(
-            "SYS-001".parse().unwrap(),
-            "Parent requirement".to_string(),
-            String::new(),
-        );
-        let parent_uuid = parent_req.uuid();
-        tree.insert(parent_req).unwrap();
-
-        let child_req = Requirement::new(
-            "USR-001".parse().unwrap(),
-            "Child requirement".to_string(),
-            String::new(),
-        );
-        let child_uuid = child_req.uuid();
-        tree.insert(child_req).unwrap();
-
-        // This should succeed
-        let result =
-            tree.upsert_parent_link(child_uuid, parent_uuid, "test-fingerprint".to_string());
-
-        assert!(
-            result.is_ok(),
-            "Expected success when both requirements exist"
-        );
-        assert!(
-            !result.unwrap(),
-            "Should return false for new link (not a replacement)"
-        );
-
-        // Verify the link was created
-        let parents = tree.parents(child_uuid);
-        assert_eq!(parents.len(), 1, "Child should have one parent");
-        assert_eq!(parents[0].0, parent_uuid, "Parent UUID should match");
-    }
-
-    #[test]
-    fn test_detect_cycles_finds_direct_cycle() {
-        let mut tree = Tree::default();
-
-        // Create two requirements
-        let req_a = Requirement::new("USR-001".parse().unwrap(), "A".to_string(), String::new());
-        let uuid_a = req_a.uuid();
-        tree.insert(req_a).unwrap();
-
-        let req_b = Requirement::new("USR-002".parse().unwrap(), "B".to_string(), String::new());
-        let uuid_b = req_b.uuid();
-        tree.insert(req_b).unwrap();
-
-        // Create A → B and B → A (cycle)
-        tree.upsert_parent_link(uuid_a, uuid_b, "fingerprint".to_string())
-            .unwrap();
-        tree.upsert_parent_link(uuid_b, uuid_a, "fingerprint".to_string())
-            .unwrap();
-
-        // Detect cycles
-        let cycles = tree.detect_cycles();
-        assert!(!cycles.is_empty(), "Should detect at least one cycle");
-        assert_eq!(cycles.len(), 1, "Should detect exactly one cycle (A→B→A)");
-    }
-
-    #[test]
-    fn test_check_would_create_cycle_detects_prevention() {
-        let mut tree = Tree::default();
-
-        // Create three requirements: A → B → C
-        let req_a = Requirement::new("USR-001".parse().unwrap(), "A".to_string(), String::new());
-        let uuid_a = req_a.uuid();
-        tree.insert(req_a).unwrap();
-
-        let req_b = Requirement::new("USR-002".parse().unwrap(), "B".to_string(), String::new());
-        let uuid_b = req_b.uuid();
-        tree.insert(req_b).unwrap();
-
-        let req_c = Requirement::new("USR-003".parse().unwrap(), "C".to_string(), String::new());
-        let uuid_c = req_c.uuid();
-        tree.insert(req_c).unwrap();
-
-        let req_d = Requirement::new("USR-004".parse().unwrap(), "D".to_string(), String::new());
-        let uuid_d = req_d.uuid();
-        tree.insert(req_d).unwrap();
-
-        // Create A → B and B → C
-        tree.upsert_parent_link(uuid_a, uuid_b, "fingerprint".to_string())
-            .unwrap();
-        tree.upsert_parent_link(uuid_b, uuid_c, "fingerprint".to_string())
-            .unwrap();
-
-        // Attempting to create C → A should fail (would create A→B→C→A cycle)
-        let result = tree.check_would_create_cycle(uuid_c, uuid_a);
-        assert!(
-            result.is_err(),
-            "Should detect that C→A would create a cycle"
-        );
-
-        // Attempting to create C → B should fail (would create B→C→B cycle)
-        let result = tree.check_would_create_cycle(uuid_c, uuid_b);
-        assert!(
-            result.is_err(),
-            "Should detect that C→B would create a cycle"
-        );
-
-        // Attempting to create D → A should succeed (no cycle)
-        let result = tree.check_would_create_cycle(uuid_d, uuid_a);
-        assert!(
-            result.is_ok(),
-            "Should allow D→A since it doesn't create a cycle"
-        );
-    }
-
-    #[test]
-    fn test_check_would_create_cycle_simple_reverse() {
-        // Test the specific scenario from the review comment:
-        // B → A exists (B depends on A)
-        // Try to add A → B (would create a cycle)
-        // This should FAIL but currently SUCCEEDS due to wrong traversal direction
-        let mut tree = Tree::default();
-
-        let req_a = Requirement::new("USR-001".parse().unwrap(), "A".to_string(), String::new());
-        let uuid_a = req_a.uuid();
-        tree.insert(req_a).unwrap();
-
-        let req_b = Requirement::new("USR-002".parse().unwrap(), "B".to_string(), String::new());
-        let uuid_b = req_b.uuid();
-        tree.insert(req_b).unwrap();
-
-        // Create B → A (B depends on A)
-        tree.upsert_parent_link(uuid_b, uuid_a, "fp".to_string())
-            .unwrap();
-
-        // Now try to add A → B (which would create cycle A → B → A)
-        let result = tree.check_would_create_cycle(uuid_a, uuid_b);
-        assert!(
-            result.is_err(),
-            "Should detect that A→B would create cycle (B→A already exists)"
-        );
-    }
-
-    #[test]
-    fn test_check_would_create_cycle_detects_ancestor_to_descendant() {
-        let mut tree = Tree::default();
-
-        // Create chain: A → B → C → D
-        let req_a = Requirement::new("USR-001".parse().unwrap(), "A".to_string(), String::new());
-        let uuid_a = req_a.uuid();
-        tree.insert(req_a).unwrap();
-
-        let req_b = Requirement::new("USR-002".parse().unwrap(), "B".to_string(), String::new());
-        let uuid_b = req_b.uuid();
-        tree.insert(req_b).unwrap();
-
-        let req_c = Requirement::new("USR-003".parse().unwrap(), "C".to_string(), String::new());
-        let uuid_c = req_c.uuid();
-        tree.insert(req_c).unwrap();
-
-        let req_d = Requirement::new("USR-004".parse().unwrap(), "D".to_string(), String::new());
-        let uuid_d = req_d.uuid();
-        tree.insert(req_d).unwrap();
-
-        // Create A → B, B → C, C → D
-        tree.upsert_parent_link(uuid_a, uuid_b, "fp".to_string())
-            .unwrap();
-        tree.upsert_parent_link(uuid_b, uuid_c, "fp".to_string())
-            .unwrap();
-        tree.upsert_parent_link(uuid_c, uuid_d, "fp".to_string())
-            .unwrap();
-
-        // This test case specifically targets the bug in can_reach:
-        // Trying to link D → A should fail because A already reaches D
-        let result = tree.check_would_create_cycle(uuid_d, uuid_a);
-        assert!(
-            result.is_err(),
-            "Should detect that D→A would create cycle (A→B→C→D→A)"
-        );
-    }
-
-    #[test]
-    fn cycles_equal_detects_wrapped_rotations() {
-        fn cycle(ids: &[&str]) -> Vec<Hrid> {
-            ids.iter().map(|s| s.parse().unwrap()).collect()
-        }
-
-        // The same cycle recorded from different entry points: rotations that
-        // wrap around the start must compare equal.
-        let from_a = cycle(&["USR-001", "USR-002", "USR-003", "USR-001"]);
-        let from_b = cycle(&["USR-002", "USR-003", "USR-001", "USR-002"]);
-        let from_c = cycle(&["USR-003", "USR-001", "USR-002", "USR-003"]);
-        assert!(Tree::cycles_equal(&from_a, &from_b));
-        assert!(Tree::cycles_equal(&from_a, &from_c));
-
-        // Same length and membership but opposite direction: not equal.
-        let reversed = cycle(&["USR-001", "USR-003", "USR-002", "USR-001"]);
-        assert!(!Tree::cycles_equal(&from_a, &reversed));
-
-        // Self-loops.
-        assert!(Tree::cycles_equal(
-            &cycle(&["USR-001", "USR-001"]),
-            &cycle(&["USR-001", "USR-001"])
-        ));
-        assert!(!Tree::cycles_equal(
-            &cycle(&["USR-001", "USR-001"]),
-            &cycle(&["USR-002", "USR-002"])
-        ));
-    }
-
-    #[test]
-    fn detect_cycles_survives_deep_chains() {
-        // A linear parent chain deep enough to overflow the call stack if
-        // cycle detection were recursive.
-        const DEPTH: usize = 20_000;
-
-        let mut tree = Tree::default();
-        let mut uuids = Vec::with_capacity(DEPTH);
-        for i in 1..=DEPTH {
-            let req = Requirement::new(
-                format!("USR-{i}").parse().unwrap(),
-                String::new(),
-                String::new(),
-            );
-            uuids.push(req.uuid());
-            tree.insert(req).unwrap();
-        }
-        for pair in uuids.windows(2) {
-            tree.upsert_parent_link(pair[0], pair[1], "fp".to_string())
-                .unwrap();
-        }
-
-        assert!(tree.detect_cycles().is_empty());
     }
 }
