+++ req-core/src/storage/markdown/frontmatter.rs
@@ -1,0 +1,164 @@
+//! YAML frontmatter (de)serialization for requirement files.
+
+use std::collections::BTreeSet;
+
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+use crate::domain::Hrid;
+
+/// The YAML frontmatter block of a serialized requirement.
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+#[serde(from = "FrontMatterVersion")]
+#[serde(into = "FrontMatterVersion")]
+pub(super) struct FrontMatter {
+    pub(super) uuid: Uuid,
+    pub(super) created: DateTime<Utc>,
+    pub(super) tags: BTreeSet<String>,
+    pub(super) parents: Vec<Parent>,
+}
+
+/// A parent requirement reference in the serialized format.
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
+pub struct Parent {
+    pub(super) uuid: Uuid,
+    pub(super) fingerprint: String,
+    #[serde(
+        serialize_with = "hrid_as_string",
+        deserialize_with = "hrid_from_string"
+    )]
+    pub(super) hrid: Hrid,
+}
+
+/// Serialize an HRID as a string.
+///
+/// # Errors
+///
+/// Returns an error if serialization fails.
+pub fn hrid_as_string<S>(hrid: &Hrid, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: serde::Serializer,
+{
+    // Use default 3-digit formatting for frontmatter serialization
+    serializer.serialize_str(&hrid.display(3).to_string())
+}
+
+/// Deserialize an HRID from a string.
+///
+/// # Errors
+///
+/// Returns an error if the string cannot be parsed as a valid HRID.
+pub fn hrid_from_string<'de, D>(deserializer: D) -> Result<Hrid, D::Error>
+where
+    D: serde::Deserializer<'de>,
+{
+    let s = String::deserialize(deserializer)?;
+    Hrid::try_from(s.as_str()).map_err(serde::de::Error::custom)
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(tag = "_version")]
+enum FrontMatterVersion {
+    #[serde(rename = "1")]
+    V1 {
+        uuid: Uuid,
+        created: DateTime<Utc>,
+        #[serde(default, skip_serializing_if = "BTreeSet::is_empty")]
+        tags: BTreeSet<String>,
+        #[serde(default, skip_serializing_if = "Vec::is_empty")]
+        parents: Vec<Parent>,
+    },
+}
+
+impl From<FrontMatterVersion> for FrontMatter {
+    fn from(version: FrontMatterVersion) -> Self {
+        match version {
+            FrontMatterVersion::V1 {
+                uuid,
+                created,
+                tags,
+                parents,
+            } => Self {
+                uuid,
+                created,
+                tags,
+                parents,
+            },
+        }
+    }
+}
+
+impl From<FrontMatter> for FrontMatterVersion {
+    fn from(front_matter: FrontMatter) -> Self {
+        let FrontMatter {
+            uuid,
+            created,
+            tags,
+            parents,
+        } = front_matter;
+        Self::V1 {
+            uuid,
+            created,
+            tags,
+            parents,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::num::NonZeroUsize;
+
+    use chrono::TimeZone;
+
+    use super::*;
+    use crate::domain::hrid::KindString;
+
+    fn req_hrid() -> Hrid {
+        Hrid::new(
+            KindString::new("REQ".to_string()).unwrap(),
+            NonZeroUsize::new(1).unwrap(),
+        )
+    }
+
+    #[test]
+    fn frontmatter_version_conversion() {
+        let uuid = Uuid::parse_str("12b3f5c5-b1a8-4aa8-a882-20ff1c2aab53").unwrap();
+        let created = Utc.with_ymd_and_hms(2025, 7, 14, 7, 15, 0).unwrap();
+        let tags = BTreeSet::from(["tag1".to_owned()]);
+        let parents = vec![Parent {
+            uuid: Uuid::parse_str("550e8400-e29b-41d4-a716-446655440000").unwrap(),
+            fingerprint: "fp1".to_string(),
+            hrid: req_hrid(),
+        }];
+
+        let frontmatter = FrontMatter {
+            uuid,
+            created,
+            tags,
+            parents,
+        };
+        let version: FrontMatterVersion = frontmatter.clone().into();
+        let back_to_frontmatter: FrontMatter = version.into();
+
+        assert_eq!(frontmatter, back_to_frontmatter);
+    }
+
+    #[test]
+    fn parent_creation() {
+        let uuid = Uuid::parse_str("12b3f5c5-b1a8-4aa8-a882-20ff1c2aab53").unwrap();
+        let fingerprint = "test-fingerprint".to_string();
+        let hrid = req_hrid();
+
+        let parent = Parent {
+            uuid,
+            fingerprint: fingerprint.clone(),
+            hrid: hrid.clone(),
+        };
+
+        assert_eq!(parent.uuid, uuid);
+        assert_eq!(parent.fingerprint, fingerprint);
+        assert_eq!(parent.hrid, hrid);
+    }
+}
